Refactor executor
Move execution functions into dedicated Executor class
